- add support for flashing to DFU
git-svn-id: https://svn.openpcd.org:2342/trunk@210 6dc7ffe9-61d6-0310-9af1-9938baff3ed1

+++ firmware/src/dfu/dfu.h
@@ -71,7 +71,7 @@ struct dfuapi {
 	int  (*dfu_ep0_handler)(u_int8_t req_type, u_int8_t req,
 				     u_int16_t val, u_int16_t len);
 	void (*dfu_switch)(void);
-	u_int8_t *dfu_state;
+	u_int32_t *dfu_state;
 	const struct usb_device_descriptor *dfu_dev_descriptor;
 	const struct _dfu_desc *dfu_cfg_descriptor;
 };
